fix: resolve CI backend test failures
- Add missing httpx dependency for CI environment
- Extend crawl4ai mocks to include deep_crawling module classes
- Add pytest conftest.py for better CI environment isolation
- Mock all required crawl4ai classes: BFSDeepCrawlStrategy, DFSDeepCrawlStrategy, etc.
- Fix import errors that were causing CI test failures

All 420 backend tests now run successfully in CI environment.
