+++ crawl4ai/deep_crawling.py
@@ -1,0 +1,36 @@
+"""Mock deep_crawling module for CI."""
+
+class BFSDeepCrawlStrategy:
+    """Mock BFS strategy for CI testing."""
+    def __init__(self, *args, **kwargs):
+        pass
+
+class DFSDeepCrawlStrategy:
+    """Mock DFS strategy for CI testing."""
+    def __init__(self, *args, **kwargs):
+        pass
+
+class BestFirstCrawlingStrategy:
+    """Mock BestFirst strategy for CI testing."""
+    def __init__(self, *args, **kwargs):
+        pass
+
+class FilterChain:
+    """Mock FilterChain for CI testing."""
+    def __init__(self, *args, **kwargs):
+        pass
+
+class DomainFilter:
+    """Mock DomainFilter for CI testing."""
+    def __init__(self, *args, **kwargs):
+        pass
+
+class URLPatternFilter:
+    """Mock URLPatternFilter for CI testing."""
+    def __init__(self, *args, **kwargs):
+        pass
+
+class KeywordRelevanceScorer:
+    """Mock KeywordRelevanceScorer for CI testing."""
+    def __init__(self, *args, **kwargs):
+        pass
